composer: Fix improper version constraint for Apache Thrift
The previous syntax of `^0.10.0` used the `^` indicator, which will
permit versions up to the next major version bump. However this is only
true for major versions >= 1; for `v0.x.y` versions, it only permits
upgrades within the same _minor_ version. This is reasonable for
dependencies that follow semver, but Apache Thrift does not follow
semver.

The new syntax of `>=0.10.0` achieves what the old syntax intended to
do, which is to allow upgrades up to the next major version.